Working version. Pull for Release.

* commit 'a9f00004f099abf6088ad20d689855ca311fa7a2':
  Refined Controller input so that menus can work. Key's now have states, and to get the state you call a function that is specific to each state. For example, to get a key down press, you call Controller::GetKeyDown(Key k);
  Added Right and Left Click detection to the controller
  Controller refactor completed, tested and working. This is a release version of the engine.
  Controller has been refactored to use the new input system. It still needs to be tested and corrected.
  Rebuilt Controller input, untested
  Refactor finished, back to working order. Pushing both to save state and push to home.
  Refactored for new class changes. There is an Error in the ID system for the GameObject2D that needs to be looked into. Pushing for remote work.
  Push to work remote
  Refactored all std::map set ups to use insert instead of [ ], updated singleton implemenation to use Class::Instance-> call instead of storing a pointer as a member.
  daily push
  Minor refactoring in the world and game object to better server the engine overall.
  Refactored Killer Physics and Killer Math to be part of the Killer Engine directly, but in their own namespaces.
  Refined a bit, routine push
  Alpha Blending now enabled when textures are rendered.
  Commit to work remotely
  Last one needed for project configuration. This one should work like a dream.
  Last try for the night

+++ Implementations/Font.cpp
@@ -131,7 +131,7 @@ namespace KillerEngine
 		//Make ID avilable first
  		id.erase(id.begin(), id.begin() + id.find_first_of("=")+1);
  		
- 		U32 charID = std::stoi(id);
+ 		U32 charID = std::stoul(id);
 
  		_fontCharData.insert(std::pair<U32, CharacterData*>(charID, new CharacterData()));
